v1.82.5

+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/GetCommentsI.kt
@@ -46,39 +46,39 @@ data class GetCommentsI (
 
     /* The comment id. */
     @SerializedName("parent_id")
-    val parentId: kotlin.Double? = null,
+    val parentId: kotlin.Int? = null,
 
     /* The post id. */
     @SerializedName("post_id")
-    val postId: kotlin.Double? = null,
+    val postId: kotlin.Int? = null,
 
     @SerializedName("creator_username")
     val creatorUsername: kotlin.String? = null,
 
     /* The person id. */
     @SerializedName("creator_id")
-    val creatorId: kotlin.Double? = null,
+    val creatorId: kotlin.Int? = null,
 
     @SerializedName("community_name")
     val communityName: kotlin.String? = null,
 
     /* The community id. */
     @SerializedName("community_id")
-    val communityId: kotlin.Double? = null,
+    val communityId: kotlin.Int? = null,
 
     @SerializedName("limit")
-    val limit: kotlin.Double? = null,
+    val limit: kotlin.Int? = null,
 
     /* To get the next or previous page, pass this string unchanged as `page_cursor` in a new request to the same endpoint.  Do not attempt to parse or modify the cursor string. The format is internal and may change in minor Lemmy versions. */
     @SerializedName("page_cursor")
     val pageCursor: kotlin.String? = null,
 
     @SerializedName("max_depth")
-    val maxDepth: kotlin.Double? = null,
+    val maxDepth: kotlin.Int? = null,
 
     /* Filter to within a given time range, in seconds. IE 60 would give results for the past minute. */
     @SerializedName("time_range_seconds")
-    val timeRangeSeconds: kotlin.Double? = null,
+    val timeRangeSeconds: kotlin.Int? = null,
 
     @SerializedName("sort")
     val sort: CommentSortType? = null,
